Introduce total_wrap_around configuration option
Allow for the detection of the wrap around of a `contains_total` field and adjust the new value accordingly.

Before this commit `current_total` being smaller than `previous_total` would result in the field being set to `None` and it would take another packet to actually calculate the difference again. This can result in some data being "lost", because the data right after the wrap around is essentially discarded.
Setting the new `total_wrap_around` option to `True` will result in `current_total` being used as the return value right away.

The option is set to `False` by default, so you have to explicitly opt-in.

Some examples:

- The `yearlyrainin` of some Ecowitt/FineOffset stations wraps around every year
- The `lightning_num` of some Ecowitt/FineOffset stations wraps around every day